requirements.txt
attrs==16.3.0
requests==2.13.0
slimit==0.8.1
lxml==3.7.1
